1.5 - Logic bugs fixing...

+++ src/main/java/api/rest/SeasFit/dto/OrderAdminDTO.java
@@ -17,7 +17,7 @@ public class OrderAdminDTO {
     private String paymentMethod;
     private String voucherCode;
     private String note;
-    private BigDecimal totalAmount;
+    private Integer totalAmount;
     private BigDecimal discountAmount;
     private int shippingFee;
     private String cancelReason;
